Change single-player game to allow multiple SETs.
In preparation for integration with multiplayer architecture. Done by
changing functions to accept SET as an argument instead of referring to the
global SET. There is a new global, SETS, containing a list of sets. The
player's set is SETS[0]. The opponent's set will be SETS[1].

Global known_best_paths was moved into SET to allow for multiple copies.