Removed most references to millis() outside of ptd.js.
Would like to centralize timekeeping to allow for fast-forwarding of game
state. Creep, creep_wave, and weapon last property were initialized to
millis(). Changed these to initiailize to set.now, which is updated by the
StatusUpdater.

This change caused the creep wave timer to initialize to NaN. This was fixed
by initializing SET.now to millis() in default_set().

References to millis() in the pause mode functionality in ui_modes.js were
not changed. Since the StatusUpdater will not run when the game is paused,
it seemed unneccesary / dangerous to change this.