Create a new copy of game source for the server.
This will need to have all attemps to draw and use the features of Processing
stripped out. However, since the objects in SET will be passed directly to the
client, those object methods should remain the same!

Unanswered question: is it OK to refer to Processing methods if they are never
actually called on the server side?