Merge PTD files into ptd.js.
Since the original code is written to all occupy the same namespace, turning
this into a system of modules would be a big job. Took the ugly but quick
option of stuffing it all into ptd.js.

node game.js executes and imports ptd.js successfully. The interface to PTD is
through ptd.start_tower_defense().